# Improve performance of `dpnp.nanmedian`